Split all the PURL functionality into a separate module
