+++ cargo-cyclonedx/src/lib.rs
@@ -20,6 +20,7 @@ pub mod config;
 pub mod format;
 pub mod generator;
 pub mod platform;
+pub mod purl;
 pub mod urlencode;
 
 pub use crate::generator::*;
